Перенес логику из activity в отдельные классы presenter
