+++ app/src/main/AndroidManifest.xml
@@ -17,13 +17,13 @@
         android:theme="@style/Theme.SamSebeTrener"
         tools:targetApi="31">
         <activity
-            android:name=".MenuActivity"
+            android:name=".activities.MenuActivity"
             android:exported="false" />
         <activity
-            android:name=".AuthActivity"
+            android:name=".activities.AuthActivity"
             android:exported="false" />
         <activity
-            android:name=".MainActivity"
+            android:name=".activities.RegActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
