--- app/src/main/java/com/example/samsebetrener/MainActivity.kt
@@ -1,50 +1,0 @@
-package com.example.samsebetrener
-
-import android.content.Intent
-import android.os.Bundle
-import android.widget.Button
-import android.widget.EditText
-import android.widget.TextView
-import android.widget.Toast
-import androidx.activity.enableEdgeToEdge
-import androidx.appcompat.app.AppCompatActivity
-import androidx.core.view.ViewCompat
-import androidx.core.view.WindowInsetsCompat
-import com.example.samsebetrener.utils.initialActivity
-
-class MainActivity : AppCompatActivity() {
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        initialActivity(R.layout.activity_main, R.id.main)
-
-        val userLogin: EditText = findViewById(R.id.user_login)
-        val userEmail: EditText = findViewById(R.id.user_email)
-        val userPass: EditText = findViewById(R.id.user_pass)
-        val button: Button = findViewById(R.id.button_reg)
-        val linkToAuth: TextView = findViewById(R.id.link_to_auth)
-
-        linkToAuth.setOnClickListener {
-            val intent = Intent(this, AuthActivity::class.java)
-            startActivity(intent)
-        }
-
-        button.setOnClickListener {
-            val login = userLogin.text.toString().trim()
-            val email = userEmail.text.toString().trim()
-            val pass = userPass.text.toString().trim()
-
-            if(login == "" || email == "" || pass == "")
-                Toast.makeText(this, "Не все поля заполнены", Toast.LENGTH_LONG).show()
-            else {
-                val user = User(login = login, email = email, pass = pass)
-
-                val db = DbHelper(context = this, factory = null)
-                db.addUser(user)
-                Toast.makeText(this, "Пользователь $login добавлен", Toast.LENGTH_LONG).show()
-                userLogin.text.clear()
-                userEmail.text.clear()
-                userPass.text.clear()
-            }
-        }
-    }
-}
