+++ app/src/main/java/com/example/samsebetrener/models/User.kt
@@ -1,4 +1,4 @@
-package com.example.samsebetrener
+package com.example.samsebetrener.models
 
 class User(val login: String, val email: String, val pass: String) {
 }
