+++ app/src/main/java/com/example/samsebetrener/presenters/MenuPresenter.kt
@@ -1,0 +1,4 @@
+package com.example.samsebetrener.presenters
+
+class MenuPresenter {
+}
