+++ app/src/main/res/layout/activity_auth.xml
@@ -7,7 +7,7 @@
     android:layout_height="match_parent"
     android:background="@color/black"
     android:orientation="vertical"
-    tools:context=".AuthActivity">
+    tools:context=".activities.AuthActivity">
 
 
     <View
