+++ app/src/main/res/layout/activity_menu.xml
@@ -5,6 +5,6 @@
     android:id="@+id/main"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".MenuActivity">
+    tools:context=".activities.MenuActivity">
 
 </androidx.constraintlayout.widget.ConstraintLayout>
