+++ app/src/main/res/layout/activity_reg.xml
@@ -7,7 +7,7 @@
     android:layout_height="match_parent"
     android:background="@color/black"
     android:orientation="vertical"
-    tools:context=".MainActivity">
+    tools:context=".activities.RegActivity">
 
 
     <View
